+++ Common/MessageCoder/MessageCoder.h
@@ -38,8 +38,8 @@ class MessageCoder
         static void CreateRoleMessage(const ROLE_ENUM role, const std::string & id, char *dest);
         static void CreateTimeToSendResultToClientsMessage(char *dest);
         static void CreateTimeoutMessage(char *dest);
-        static void CreateStartAnalyseWork(const unsigned short respTime , const Board & board, const std::string & id, const std::string & jobId, const bool isFirstWorker, char *dest);
-        static void CreateStartAnalyseWorkAndReturnNResultFast(const unsigned short respTime, const unsigned int numOfResultToReturnFast, const Board & board, const std::string & id, const std::string & jobId, char *dest);
+        static void CreateStartAnalyseWork(const unsigned long long respTime , const Board & board, const std::string & id, const std::string & jobId, const bool isFirstWorker, char *dest);
+        static void CreateStartAnalyseWorkAndReturnNResultFast(const unsigned long long respTime, const unsigned int numOfResultToReturnFast, const Board & board, const std::string & id, const std::string & jobId, char *dest);
         static void CreateOkMessage(const std::string & id, char *dest);
         static void CreateNotOkMessage(const std::string & id, NOT_OK_REASON reason, char *dest);
